make a cosivina version of Dinasaur
This is to be able to use the oriented-object paradigm, which will make simpler the addition of new features.
(Having everything in one function would have become very messy).
However, Cosivina is really slow. Probably cause of the Convolution.
Try to remove the inf cutt-off in GaussKernel1D set-up. If not working try to have a weight map instead of convolution.